Specify package version in __init__.py only.
Now that we are adding sphinx documentation, the version of the package
appears in more than one place in the repository (setup.py and
doc/source/conf.py). Instead of having to maintain the version in these
multiple places, we define the version in cvxpy/__init__.py only. The
other locations import the package, and take the value from
cvxpy/__init__.py.