Add example notebook sparse_covariance_est from CVX.
This example was ported from the CVX example "Sparse covariance
estimation for Gaussian variables" by Joelle Skaf, 04/24/2008.

Porting was done by Judson Wilson, 05/22/2014.

Overall, code is similiar, besides the port changes, although several
issues were fixed. The original problem thresholded elements of the result
to be 0 if they were less than 1e-4. The correct way is to threshold
the absolute value of the element. After that, the value of alpha
had to be greatly reduced to cause true sparsity, since all the
thresholded values were large-magnitude negative values.

As an improvement, plots for 3 different values of alpha were added.