Add example notebook clock_mesh from CVX.
This example was ported from the CVX example "Sizing of clock
meshes" by Lieven Vanderberghe and others.

Porting was done by Judson Wilson, 05/26/2014.

Overall, code is very similiar, almost a straight port.
Presently execution is about 100 times slower than on CVX, so
I had to reduce the number of test points, npoints, by a factor
of 2, and it still takes forevor. Also, the solutions are a
little bit different, and the constraints aren't exactly met.

Also made some slight cosmetic tweaks to the text on the graphs.