Add example notebook nonneg_matrix_fact from CVX.
This example was ported from the CVX example "Nonnegative matrix
factorization" by Argyris Zymnis, Joelle Skaf, and Stephen Boyd.

Porting was done by Judson Wilson, 6/2/2014.

Generally this a direct port, with modifications necessary for the
port. For some reason CVXPY will not converge as quickly as
CVX. In fact its possible that the solution will never converge.
I am not sure if this is true. The reason is unclear. It may
be an artifact of the data and the stopping criterion of the solver
at each iteration.