Performance improvements for formatting constraints
- scipy.block_diag is slow, because it converts matrices to COO and
  has an expensive nested for loop; this commit removes a call to block_diag
  and implements an expensive matmul more efficiently in
  conic_solver:format_constraints
- applying parameters in ParamConeProg can be slow, even when there are no
  parameters; multipying a large SciPy sparse matrix by [[1]] on the right
  can take 2 seconds. this change adds a fastpath for problems with no parameters

Note that the slowness that this commit addresses was seen even for problems
without parameters.

Before this change, the TV inpainting benchmark in cvxpy/tests/test_benchmarks.py
would be killed by the OS; now it takes 17s (compared to 2s at master). With
this commit, solving an inpainting problem with dimensions (52, 52, 1) takes
0.2s (compared to 0.04s at master). Without this commit, solving the same
problem took 44s, for a speed-up of ~200x

Much of the remaining slowness can be attributed to cvxcore.