Implement and test DPP grammar
This change (re-)implements DPP. The implementation consists of two key parts:

(1) a context manager (treat_params_as_affine) is added that marks parameters
    as non-const; this makes them affine, like variables
(2) the curvature of the product atom (MulExpression) now depends upon whether
    DPP analysis is underway. if a treat_params_as_affine context manager is
    active, then additional restrictions are added for a product to be convex
    (affine). Specifically, at most one of the arguments can contain parameters,
    and the argument containing parameters must be affine.

The implementation of is_dpp(self) is then just:

```
    with treat_params_as_affine(self):
        return self.is_dcp()
```

Notably, this means multiplying parameters is not allowed under DPP.
This change disables the transformation that automatically substituted
variables for parameters in products of parameters.

Problems that are DCP but not DPP can be solved; however, a warning is
printed when a user tries to solve a DCP-but-not-DPP problem.

With this change, the matrix stuffing reductions require problems to be DPP
(since otherwise cvxcore's behavior is undefined).

This change also adds unit tests for DPP.

All CVXPY tests are passing (except for the two performance tests).