Sync up cplex_conif.py with cplex_qpif.py
This commit synchronizes the cplex_conif.py file with
cplex_qpif.py. In particular, we delay querying the solution/dual
values until we get to the CPLEX.invert() method (and only if the
solution status is in s.SOLUTION_PRESENT). It also simplifies the
code. The primary reason for doing this was to fix the failing
test_atom.py:test_partial_optimize_special_var unit test.