Fix cplex 201911 (#869)
* Minor cleanup in scs_conif.py

Use constants instead of hard-coded values in dims_to_solver_dict().

* Sync up cplex_conif.py with cplex_qpif.py

This commit synchronizes the cplex_conif.py file with
cplex_qpif.py. In particular, we delay querying the solution/dual
values until we get to the CPLEX.invert() method (and only if the
solution status is in s.SOLUTION_PRESENT). It also simplifies the
code. The primary reason for doing this was to fix the failing
test_atom.py:test_partial_optimize_special_var unit test.

* Share code that maps CPLEX status codes to CVXPY status codes

* Share code that hides CPLEX output between conic and qp interfaces

* Share code that sets CPLEX parameters between conic and qp interfaces

* Replace stale TODO comment in cplex_conif.py

* Handle solution status codes that will be removed in CPLEX 12.10

With this commit all tests pass with CPLEX 12.10 beta.

* Remove kwargs removal code in cplex_conif.py:set_parameters

This addresses a reviewer comment. The offending code was here before
any changes on this branch, but they looked out of place to me as
well.

* Add a big comment about dual values in cplex_conif.py:invert

This addresses a reviewer comment.