DQCP: Fix detection of monotone real functions
Fixes a bug in DQCP analysis in analyzing monotone real functions; DQCP
analysis previously did not check whether the argument of a scalar function
was also scalar. This bug caused DQCP analysis to be too permissive, incorrectly
labeling things as DQCP and causing the reduction to fail in surprising ways.

As an extension, this change adds the vector atoms cp.max and cp.min to the
DQCP atom library, letting users take the max or min of quasiconvex or
quasiconcave expressions. Previously, only the elementwise cp.maximum and
cp.minimum atoms were allowed.