[DGP] Propagate Parameter names in Dgp2Dcp
Parameters in a DGP problem are converted into new parameters in a DCP
problem, via Dgp2Dcp. Before this change, the names of the original parameters
were not properly propagated to the new parameters, leading to an error when
the DCP problem was printed. Users don't interact with the output of
Dgp2Dcp directly, but this bug showed up during internal testing.