DPP for DGP (#1010)
* [DGP] DPP for DGP, all tests passing.

Implements DPP for DGP. The DPP ruleset for DGP is analogous to DPP for DCP:
the program must be DGP in both the parameters and the variables, with
an exception for the power atom (which is an analog of multiplication in DCP).
Specifically, an expression `expr**parameter` is considered DPP (log-log affine)
as long as `expr` is not parametrized. For simplicity, the parameter cannot
be replaced with a log-log affine composition of parameters, ie it must be a
raw parameter.

This change requires delaying the rational approximation to the exponent of the
power atom required by Dcp2Cone. That's now done in a dcp2cone reduction
(power_canon), instead of at initialization. This leads to a slight semantic
difference in power's monotonicty calculation, which was previously based
on the rational approximation to the exponent p. With this change, the
monotonicity depends on the original value.

* [DGP] Derivatives

Implement backward() and derivative() for DGP, with some basic tests.

* simplify is_dpp for power

* Move rational approximation to exponent p back to power atom, for DCP programs

* make compute_once decorator specialize on args, kwargs

* dpp flag for is_dcp, is_dgp

also fix bugs in dpp checking for constraints

* Test nonconvex constraints are not DPP

* More derivative, dpp tests, fix power monotonicity for params

* More DGP derivative tests

* Update docstrings in problem.py

TODO website documentation, specifically DPP for DGP problems

* DPP web documentation

* web documentation for derivatives

* doc fixes

* Lower tolerances for derivative tests

To get py35 builds passing

* Suppress some DPP warnings in tests

* Filter warnings for DPP in bisection, add TODOs

* two more dpp tests

* doc fixes for reshape

* pf eigenvalue is increasing in entries

* update test

* dpp for gmatmul

* fix gmatmul dpp bug

* wip speed up for param apply

* wip

* fixes for reduced problem data tensor construction

* fast path for apply_param_jac

* Address review comments

* Fix error in power sign, which was introduced during refactoring