Cache auxiliary output for some solvers
We currently cache the OSQP/SCS results dicts in problem._solver_cache. This
is useful when debugging a call to solve. This
change also caches the ECOS results dict and the entire Gurobi model;
this makes it possible to inspect the solver's performance
programmatically, without looking at the verbose output.

We should implement similar changes for all solvers, in a future change. It might be
a good idea to expose cached solver information in a public API in a
future PR.