Cache auxiliary output for some solvers (#1062)
* Cache auxiliary output for some solvers

We currently cache the OSQP/SCS results dicts in problem._solver_cache. This
is useful when debugging a call to solve. This
change also caches the ECOS results dict and the entire Gurobi model;
this makes it possible to inspect the solver's performance
programmatically, without looking at the verbose output.

We should implement similar changes for all solvers, in a future change. It might be
a good idea to expose cached solver information in a public API in a
future PR.

* Add extra_stats to SolverStats

Add an extra_stats attribute to SolverStats, to store information
returned directly by the solver. Populate this attribute for ECOS, SCS,
OSQP, and GUROBI; other solvers can be added later.

Co-authored-by: Akshay Agrawal <akshaya@netflix.com>