Big data bug fixes (#1074)
* Fixed the overflow error for the conic solver

I fixed an overflow error which was preventing the conic solver from solving large optimization problems. I am attempting to do the same thing for the MOSEK solver

* Reversing MOSEK changes to pull only conic solver fixes

I'm reverting my MOSEK solver fixes so that I can the only thing that gets pulled into the main branch is the conic solver overflow fix. This means this branch fixes an overflow error that was occuring on the shape check of the matrix. the datatypes of the variables were int32 which was causing the error. Changing these datatypes to int64 (the edits I made) allows larger optimization problems to be solved and no errors have been encountered.

Co-authored-by: SpencerStingleyUNCG <srstingley@uncg.edu>