Reversing MOSEK changes to pull only conic solver fixes
I'm reverting my MOSEK solver fixes so that I can the only thing that gets pulled into the main branch is the conic solver overflow fix. This means this branch fixes an overflow error that was occuring on the shape check of the matrix. the datatypes of the variables were int32 which was causing the error. Changing these datatypes to int64 (the edits I made) allows larger optimization problems to be solved and no errors have been encountered.