Fix handling of degenerate problems
The case-check at line 121 was meant to identify when a problem had no conic constraints. However, it actually identified when a problem had more linear equations than variables (in which case the constraints are either redundant or the problem is trivially infeasible). This change should get us the behavior that was intended from the start.