Verbose logging specific to compilation
Adds a keyword argument `verbose` to the solve method. When
verbose=True, CVXPY will log information related to both problem
compilation and solving.

For beginners, the distinction between compilation and solving is not
clear; this can lead some users to conclude that the solver is taking
too long, when compilation was the bottleneck (or vice versa). This
logging is meant to help alleviate that confusion.

For beginners and advanced users alike, when compiling very large
problems, it can be helpful to have logging output describing the
various stages of compilation and noting when they finish.