Validate constraints on problem creation
Problem canonicalization fails silently when an expression is included
in a problem's list of constraints.

This change modifies the Problem constructor to check that each
constraint is in fact a Constraint (or a bool)

Previously DQCP relied on duck-typing and included lazily evaluated
constraints in the constraint list. This change modifies the DQCP
reduction to store the lazy constraints elsewhere.

Fixes #1462