Fix integer overflow on Windows in TensorRepresentation.flatten_tensor (#2461)
Numpy's default integer type on Windows is int32. The implementation in
flatten_tensor explicitly converts Python integers into int64, but in
on place it does it after the damage has been done on Windows.

On Windows, `np.prod` returns an int32 when given a Python tuple of ints,
but can be directed to work with int64 instead. This prevents an integer
overflow for e.g. `np.prod((60000, 60000))`, which returns a negative
number on Windows. Setting the dtype for `np.prod` explicitly fixes that.