Build(pre-commit): Update ruff from 0.1.11 -> 0.8.1
[changelog for ruff](https://github.com/astral-sh/ruff/blob/main/CHANGELOG.md#081)

This updates ruff and also removes a deprecation warning about select
being in a `lint` section of the `pyproject.toml`

Also migrate to `extend-exclude` as the default exclude is
[pretty decent](https://docs.astral.sh/ruff/configuration/)

Also ignore [E721](https://docs.astral.sh/ruff/rules/type-comparison/)
This is a newer check and fails in many files.
Keeping it ignored mimic behavior pre-upgrade.
Another PR seems appropriate to fix `is / is not` vs `==/!=`