Fix link in dcp tutorial
In the DCP tutorial there is a link to a DCP website from stanford, but that address now hosts the 'developmental community of practice'. Also interesting but probably not what we wanted to see.

I looked into the wayback machine: https://web.archive.org/web/20210506191139/https://dcp.stanford.edu/

And I can't find the original resource anymore so I guess this is the closest to it? Unless someone knows where the new DCP tutorial is hosted