Treat assignment of numpy arrays and lists similarly (#2913)
* Add test for constraint violation with boolean assignment

This concerns #2900.

* Add test for matrix_utilities.convert method

* Add test for value and dtype after boolean and integer assignment

* Treat assignment of numpy arrays and lists similarly

This fixes an issue where lists and numpy ndarrays were treated
differently when assigned to a Leaf's value property.

When assigning a list (or numpy matrix), the provided object is first
converted to a numpy ndarray with float dtype.  However, when provided a
numpy ndarray, the object was being assigned as provided.  As a result,
lists of bools and ints were converted to numpy arrays of floats, while
numpy arrays of bools and ints retained their original dtypes.  This
would result in errors when attempting to perform certain arithmetic
operations on these arrays.

Now, we make lists and numpy ndarrays follow the same code path during
the conversion process.

This fixes #2900.