Fix sparse convolution for the Scipy backend
If the lhs operand has zeroes in it, they get consumed by the .tocoo
call. This means that the amount of nonzero elements computed beforehand
is no longer correct, yielding a dimension mismatch a few lines later.

This is a tentative attempt at fixing this, should be carefully analyzed
by someone that can actually read this numpy juggling.

Fixes #2949