Support for cvxpy.convolve into complex2real, fix sparse convolutions in Scipy (#2947)
* Add support for cvxpy.convolve into complex2real

Altough cvxpy.conv is supported, cvxpy.convolve is a different atom
and thus requires to be explicitly specified in order to be supported.

Fixes #2946

* Add a test for complex cp.conv and cp.convolve

* Fix sparse convolution for the Scipy backend

If the lhs operand has zeroes in it, they get consumed by the .tocoo
call. This means that the amount of nonzero elements computed beforehand
is no longer correct, yielding a dimension mismatch a few lines later.

This is a tentative attempt at fixing this, should be carefully analyzed
by someone that can actually read this numpy juggling.

Fixes #2949